Add script to link GISAID EpiFlu metadata and sequences
Links records downloaded from GISAID EpiFlu to output NDJSON that
can be processed in a `augur curate` pipeline. Each record represents a
single GISAID record with metadata and 10 segment sequences. From what
I've seen, the "HE" and "P3" segments are empty but they are kept for a
complete representation of the downloaded data.