link-gisaid: Add assertions for potential data issues
Fail loudly to guard against potential data issues:

1. If not all segment columns are not present, then GISAID might have
changed the format of the Excel file and the expected `SEGMENTS` or
`SEGMENT_COLUMNS` need to be updated.

2. If all records are missing segment accessions, then GISAID might
have changed the format of the segment ids and the
`SEGMENT_ACCESSION_PATTERN` needs to be updated.

3. If all records are missing segment sequences, then the GISAID Excel
file and FASTA file might be from different downloads and therefore
do not contain matching records.