Add skeleton ingest workflow
Add a skeleton ingest workflow that can start from a local pair
of GISAID Excel/FASTA files and create an output of metadata TSV
and 8 segment FASTAs. Custom scripts to be added in the subsequent
commits that will try to recreate the curations done in fauna and
split the NDJSON by subtype.