ingest: Split curated NDJSON by lineage
After curation of the NDJSON records, split the NDJSON by the
"lineage" field so that we can output the final metadata TSV and
sequence FASTAs split by lineage.

This change only includes the seasonal flu lineages, but we could
potentially expand the `GISAID_LINEAGE_MAP` in
"ingest/scripts/standard-lineage" to annotate other lineages and add
them to the expected `LINEAGES` in the "ingest/Snakefile".