ingest: Define expected lineages and segments in config
Allows users to override which lineages and segments to parse in
the ingest workflow. To avoid forcing users to redefine the metadata
columns for the provided segments, the segment related columns are
automatically added to the final metadata TSV.

Mainly motivated by the fact that my latest test data didn't include
yam lineage data and causes an error in the workflow. Rather than
modifying the workflow to handle missing lineages, just define the
expected lineages in the config. I think this better fits our needs of
a semi-automated workflow.