ingest/parse-gisaid-location: Add strain location parsing
Use the strain location if the metadata geolocation values are missing
any of the expected values. This commit just adds the strain location
as the next specific geolocation value if it does not match any of the
metadata geolocation values.

A few downsides to including the strain locations with this method:

1. The strain locations are _not_ standardized both in the spelling and
which geolocation value they represent, i.e. it could be country,
division, or location. We could potentially use pycountry's `search_fuzzy`
method to help with country and division values but not specific
locations.

2. This can result in some weird location values like

```
{
    "location": "North America / United States",
    "strain": "A/Baltimore/123/2025"
}
```
gets parsed as
```
{
    "region": "North America",
    "country": "United States",
    "division": "Baltimore",
    "location": "?"
}
```
We could potentially pull in fauna's geo_synonyms.tsv and do some kind
of merge of the geolocation values from the geo_synonyms.tsv and
the metadata geolocation values. The other option is to keep these
geolocations and just fix them with custom geolocation rules.

3. There are definitely mismatches between strain locations and the
metadata geolocations. These are currently flagged as warning logs,
but would probably require manual annotations to fix them.