Add script and template to automate narratives
Adds a script to load a Markdown template written in Jinja2 syntax [1]
and write out a narrative stub with the correct dates in the narrative
frontmatter and the Nextstrain URLs. The included Markdown template
assumes that the narrative will be written to a `narratives/path.md`
path with images in the same repo's `figures/` directory. We can run
this script from the workflow to automate narrative creation and
optionally upload the narrative stub to the desired Nextstrain Group.

Closes #163

[1] https://jinja.palletsprojects.com/en/stable/templates/