Run Nextclade in GHA for uploads
Adds a job to the upload GitHub Action to run Nextclade on all sequences
before triggering downstream workflows. This change ensures that
Nextclade annotations in S3 reflect the latest data when other workflows
run. This commit removes the option to trigger the separate Nextclade
GHA from the upload action. Thanks to @tsibley, this implementation
reuses the Nextclade GHA, so we can still manually trigger just that
action later when we need to.

Co-authored-by: John Huddleston <huddlej@gmail.com>
Co-authored-by: Thomas Sibley <tsibley@fredhutch.org>