Start prototyping haplotype assignment
Starts adding logic to annotate haplotypes in the phylogenetic workflow
from Nextclade annotations and user-defined haplotype definitions. This
is still a work in progress, since the core workflow's Nextclade
alignment doesn't provide subclade annotations or substitutions relative
to subclades that we need for haplotype assignment.

Also, we're probably going to want some additional features from the
haplotype assignment script like the ability to default the haplotype
value to the record's clade when no haplotype is defined and the
ability to annotate haplotypes on internal nodes so we get branch labels
that we can access via Auspice URLs in narratives.

Also also, we already have two different "derived haplotype" scripts and
rules in this workflow which makes this prototyped functionality more
confusing in the context of the whole code base. We should revisit the
value of all these various approaches and consolidate to fewer options.