Allow haplotypes to default to current clade
Adds an option to the haplotype assignment logic to make records without
assigned haplotypes to default to the record's clade name instead of a
hardcoded value. This option is most helpful in the context of the
phylogenetic workflow where we may not have explicitly listed all clades
in the haplotype definitions.