ingest: Add nextstrain-automation config and rules
Start the path for the semi-automated ingest workflow by adding rules
for uploading files to S3. This deviates from the pathogen-repo-guide
in order to use the wildcards for lineages and segments to define which
files to upload instead of listing out every single file in
`files_to_upload`.

Once we've completely switched over to the new ingest workflow,
we can remove the metadata/sequences upload rules from
profiles/upload/upload.smk.