ingest/prepare_ndjson: Dedup NDJSON by GISAID EPI ISL
There's a chance that the same GISAID record was included in multiple
downloads that we are concatenating, so dedup the NDJSON by the
GISAID EPI ISL.

This does the simple thing of keeping the first record and dropping
any duplicates found, so we order the globbed GISAID_PAIRS by reverse
order to prioritize the latest downloads. If the user is defining
specific pairs of inputs in the config, then we respect whatever order
they provide.