ingest: Add manual upload workflow
Start the path for the semi-automated ingest workflow by adding a
manual upload workflow. This is an independent Snakemake workflow that
is expected to be run manually by the Nextstrain team. The workflow will
upload pairs of GISAID metadata and sequences files to the private
AWS S3 bucket `nextstrain-data-private`.

These pairs of unprocessed GISAID files on S3 will be the starting point
for the automated ingest workflow that will be added in the following
commits.