ingest: Filter for seasonal-flu records based on lineage and host
Filter for seasonal-flu records after the lineages have been
standardized so that we can filter for the requested lineages provided
in the config. Specifically filters for "Human" hosts so that we don't
have to filter for it during the GISAID download process. This way we
can use the same central NDJSON for both seasonal-flu and avian-flu.