Run Nextclade when no annotations are available
Adds logic to run Nextclade on all input sequences when no Nextclade
annotations are available through other sources (e.g., S3). This change
fixes an issue where external users would not have been able to run the
workflow without first producing Nextclade annotations that are now
required to merge with metadata. With this change, the GISAID quickstart
in the README works as expected. This commit uses a ruleorder
declaration to prefer the S3 copy of Nextclade annotations when those
annotations are available.

As a side effect of this change, we no longer need to prepare/maintain
a Nextclade TSV for the CI build and we no longer need to create a
subset of Nextclade annotations for the emerging haplotypes rule. This
commit deletes references to those unneeded rules/files.