Update Auspice config for CI and GISAID builds
Updates the Auspice config JSON for both the CI and GISAID builds to use
the segment-specific file which has the new QC status coloring. This
change allows us to confirm that the coloring is working in our CI
builds and allows external users to get coloring by QC status without
any additional work on their part.