Remove accession from parsed GISAID sequences
Fixes an issue with the GISAID quickstart workflow where Nextclade would
output the full FASTA defline (which contains the strain name and
accession separated by a space) as the "strain" in its TSV output,
causing the downstream `augur merge` step to fail to merge those
Nextclade annotations with the parsed metadata. This commit removes the
accession from the parsed sequences completely, so the downstream merge
works as expected.