Set a default artifact name for Nextclade action
We rarely run the Nextclade GitHub Action outside the context of the
upload action these days, but when we do, we need to specify an artifact
name or the workflow will fail to validate during the "run-build" step
of the pathogen-repo-build and exit without waiting for the AWS Batch
job to finish. This commit sets a default artifact name that matches the
name used in the upload action.