Do not force include all references in full trees
Including all reference strains in full trees produces a MRCA that is
far too back in the past to be useful for this recent time view. Keep
the "include" option to prevent the local clock filter from dropping any
of these reference strains that do belong in the tree, though.