Explicitly root raw tree before pruning reference
Not all methods of tree inference will root the tree with the reference
sequence like we've assumed so far. For example, the CMAPLE algorithm in
IQ-TREE will try to find the best root in the alignment instead of using
the first sequence in the alignment as the root (default behavior for
standard IQ-TREE). This commit explicitly roots the raw tree with the
reference node before pruning the node. This approach mirrors the newer
feature in augur refine that accomplishes the same functionality.