ingest: Add custom apply-geolocation-rules
The first pass through `apply-geolocation-rules` standardizes the geolocation 
names across the board according to the default Augur rules. The second pass 
through `apply-geolocation-rules` specifically updates the region names to match 
the expected regions for seasonal flu analyses.

Doing this in a two-pass process to avoid having to maintain custom geolocation 
rules for everything in seasonal-flu. Instead, we only need to maintain the list 
of custom regions.