Remove rules and files related to fauna data workflow
Once we have fully switched to the new ingest workflow and the builds work 
as expected, we can remove a bunch of rules and files related to the fauna 
data workflow.

Note this does _not_ touch anything related to titers yet because we still 
depend on fauna for titer data.