.github/workflows/upload: switch to ingest workflow
Preparing to switch prod builds to the new ingest workflow.

Replace the sequences and metadata upload_* targets with the new ingest
workflow, including an input for triggering the avian-flu GenoFLU workflow.
`upload_titers` depends on ingest since we want to add QC steps to check
if titers have matching strains in the ingest output.

I have _not_ removed the the sequences and metadata upload_* rules in the 
Snakemake workflow so we can easily fallback to the fauna workflow if needed.