.github/workflows/upload: Add ingest workflow
Preparing to switch prod builds to the new ingest workflow.

Replace the sequences and metadata upload_* targets with the new ingest workflow.
Leaving `upload_titers` as an independent job for now since it does not 
technically require the ingest workflow to complete. In the future, we should 
add QC steps to check if titers have matching strains in the ingest output.

I have _not_ removed the the sequences and metadata upload_* rules in the 
Snakemake workflow so we can easily fallback to the fauna workflow if needed.