Update vendored references
Use the top level `shared/vendored/scripts` throughout workflows since it is 
the latest version and the `ingest/vendored` will be removed. 

Uses `workflow.current_basedir` for the paths to be relative to each *.smk file 
without having to think about how the paths will be relative to the 
main Snakefile. Based on feedback for similar changes in 
<https://github.com/nextstrain/measles/pull/62#discussion_r2101303434>