WIP: ingest-titers workflow
Start the ingest-titers workflow, beginning with CDC data.
The plan is to standardize the raw titer TSV column names and then 
run them through the same curation pipeline so that all titer data 
follow the same curation steps. 

Any CC specific steps should be done prior to the main curation rule. 
For CDC data, this is just a simple standardization of columns that is added 
in this commit