WIP: ingest-titers: copy fetch_from_fauna rule from ingest
Copied from the sequences ingest workflow
<https://github.com/nextstrain/seasonal-flu/blob/cd564882ed63cf1b791f8ac0651be69cd7cdd468/ingest/rules/curate.smk#L17-L32>

I tried to use Snakemake modules to import the rule _without_ copying it, 
but ran into a `KeyError` for the config. Even though I was only trying to use a 
single rule from the curate.smk file, the whole Snakefile gets parsed and 
requires config params that are not used in this workflow. To properly import 
I would either need to 
1. move the rule to a separate snakefile
2. rewrite the curate.smk file to _not_ require config params

Both options seem not worth it for a rule that will be eventually completely 
removed, so just copying for now.