Increase tip frequencies resolution
Monthly pivots and 1 month bandwidth was washing things out, especially at the 6m resolution that we often care about. Increasing to weekly pivots and dropping bandwidth from 1/12 to 1/20 works well empirically for 2y trees. If this to too granular for 12y trees we can add build specific parameters, but I wanted to start without this complication.