Export a distinct Nextclade subclade column
Copies the Nextclade "subclade" column to a new "subclade_nextclade"
field and adds this new column to the Auspice config JSON for H3N2 HA as
a proof of concept for exporting Nextclade-based clade annotations
alongside the `augur clades` annotations. Since the Nextclade annotation
of subclade already exists in the metadata we pass to `augur export`, we
just need to rename that column to avoid conflicting with the "subclade"
attribute that also appears in the node data JSON inputs to `augur
export`. To avoid breaking any downstream logic that relies on the
Nextclade column name of "subclade", I've chosen to duplicate the
original column to a new column with the desired name.

Related to #254
Related to #131