Handle mismatches between HA and NA clade names
Updates the logic of the workflow and the table-to-node-data script to
handle the fact that we want to select a subset of metadata/Nextclade
columns that correspond to historical clade names for HA ("clade") and
currently maintained clade names for HA ("subclade") and NA ("clade")
which have different column names, different node attribute names, and
different branch labels. In an ideal world, we might drop the historical
HA clades and standardize the names of our Nextclade, node, and branch
attributes to "clade" across all segments, but for now, we need this
additional complexity in the workflow to handle the complexity in names.