Support fitting titer models per reference virus
Allow titer collections to request a titer model run per reference virus
in each collection. The workflow aggregates the resulting model results
into a single measurements panel JSON which allows users to display
inferred titer measurements per virus in the tree in both the
measurements panel and in the tree coloring by measurements for a
reference selected from the panel. This logic attempts to recreate a
nextflu feature which allows users to click on a titer reference virus
to color the tree by the measurements against that virus and then choose
to color by the titer model fit to that virus.

Related to #214