ingest-open: Add rules to fetch from GenSpectrum
I had initially tried to use the shared `path_or_url` as input
to a trial rule to use Snakemake's remote http handling, but ran
into issues with stale data when I changed the query parameters.

I _think_ this is due to the locally saved .snakemake/storage/...
filename not including the query params, so the query param changes
did not trigger a new fetch of the data.