[ingest] type/subtype/lineage for all records
The primary way data is analysed is by first subsetting to the subtype
or lineage of interest, and so setting these values for all records will
enable a suite of analyses which were previously not possible.

Previously (for A viruses only) we sometimes set the lineage to the
subtype (e.g. h3n2, h1n2) and not for others (e.g. h3n3, h5n1). We now
set the lineage to the subtype in all cases except for h1n1/pdm09
(subtype=h1n1, lineage=h1n1pdm09). This means we expand our previous
ignoring of the upstream Lineage="seasonal" to apply to any subtype
(previously we would ignore it for some subtypes and for others they
would be assigned "TBD" type/subtype/lineage).

Because the data is messy our subtypes & lineages are inherently messy.
This means we now have subtypes of "h3", "n1" etc. In practice, if you
wanted to (e.g.) fully analyse H3 strains then you would have to subset
by many different subtypes (e.g. h3, h3n1, h3n2...), and potentially
consider the type='A' records without any subtype assigned.

There are no changes to the output seasonal-flu ndjson, but the larger
curated ndjson has many changes. Beyond the updates to 87k records with
TBD type/subtype/lineage, we also change the lineage for many records
from 'n/a' to the subtype (see above).