[dev] add script to compare NDJSONs
This is useful when making changes to the curation steps. For instance,
assuming we have stored a good copy of the seasonal-flu.ndjson created
before we made changes we can diff the new one against this.

I plan to eventually write up a recommended workflow so that anyone can
confidently make code/curation changes, and this workflow will leverage
scripts such as this.