[ingest] apply lineage annotations early
This starts the move towards applying annotations early and at the
relevant curate step rather than at the end. The benefits are that we
can provide warnings when unexpected situations arise safe in the
knowledge that there won't be a later step which corrects the record.

I've changed the format such that we now also update the vtype & subtype
which is expected to result in large number of changes to records. In
total 8,097 records are changed by this commit.

* n=8058 had their type & subtype changed from "tbd" to (e.g.) "a" &
  "h3n2". These changes were the only ones I expected when writing the
  code here

* n=14 had their subtype changed from "n/a" to either "h1n1pdm" or
  "h3n2". All of these records had raw info of "Subtype": "A",
  "Lineage": "" so these changes also make sense.

* n=25 had their subtype changed, e.g. "h3n2" → "h1n1". As an example, a
  record after curation with the previous code was:
  `{"vtype": "a", "subtype": "h3n2", "lineage": "h1n1pdm"}`
  and using this updated code is now:
  `{"vtype": "a", "subtype": "h1n1", "lineage": "h1n1pdm"}`
  These will be revisited in a subsequent commit as they flag up
  inconsistencies in the underlying annotations rather than the code
  changes introduced here.

In addition, annotations for 5 EPI ISLs are flagged as unnecessary