[dev] show script name in log output
Makes searching through the log files much easier (e.g. with `grep`) and
also makes it easier to read through the log files as there's now a
clear mapping back to the part of the curate chain where the warning
came from.

This starts a pattern I'd like to roll out to all our Augur curate
scripts, but that's for another day.