[ingest] Minor refactoring of location parsing
The shifting of the `len(split_locations) < expected_num_locations`
earlier should speed things up a little as we're no longer unnecessarily
regex-matching on the strain name

By keeping the `Lineage` field unchanged (i.e. capital L) we keep this
field in the output NDJSON which allows better troubleshooting. This
will be dropped when we do the TSV fields subsetting.