[ingest] add location fixes
The full influenza ingest now completes without any warnings.

Most changes are to non-seasonal-flu records, perhaps as expected. There
are only 168 record changes to seasonal-flu, all of which are
improvements.

As will be described in the upcomming docs commit, the process for
geographic curation is multi-layered and complex. For instance, the
`parse-gisaid-location` sets "Maricopa County AZ", which `augur curate
titlecase` chances to "Maricopa County Az", which we then change back to
"Maricopa County AZ" via a custom `geolocation_rules.tsv` entry. I'm
sure there are improvements to the process that we can make over time...