[ingest] [to revert] hide warnings re: duplicates
While the files are being fetched from fauna (and those files are
changing ~every week) it's not practical to fix the underlying issues.
We should re-enable these warnings once the files are in this repo and
we can fix them. (P.S. I don't want to keep the warnings around in the
logs as I want any warning shown to be fixable/fixed.)