[ingest] unify strain name modifications
Unifies the code across the now-diverged upload scripts in fauna.

There are no changes to seasonal-flu strain names.

Because the fauna database represents a series of uploads, each using a
specific version of the fauna upload code, applying the current code to
all (raw) records is not expected to match the database records. That
said, diffing against the (fauna) avian-flu TSV is instructive to see
the changes that will happen when we shift to this new ingest
approaches.

We are able to match 61,180 / 62,711 (98%) TSV records (matching via HA
accession, using unique strain-name matches as a fallback). Of these,
n=21 strains differ only by case and n=32,405 (53%) differ by more than
case.

Looking at the 32,405 mismatches:
* n=24,849 (77%) are  due to the replacement of dashes in older records
  on fauna (see comments in this commit's code).
* n=799 (2.5%) are identical when we remove punctuation characters (e.g.
  'a/b' is considered the same as 'a_b').
* n=191 (1%) are different only due to the new strain name having a
  /YYYY ending
* n=211 (1%) are different only due to the fauna strain having a
  trailing HxNx
* n=11 (0%) are different only by the first part (typically the type!)
* n=218 (1%) are different only by the second part (typically the host)
* n=2,932 (9%) are different only by the third part (typically the
  location, and my guess is these differences reflect code changes over
  time to update location synonyms/fixes)
* n=1,586 (5%) are different only by the fourth part (typically a strain
  ID field)
* n=10 (0%) are different only by the fifth part (typically the year)
* n=1,598 (5%) strain name changes unexplained by the above reasons