ingest: Support multiple of the same segment per record
Within GISAID, a single record can have multiple sequences for the same
segment. This commit adds support for multiple sequences of the same
segment by changing the expected schema for the "sequences" key in the
NDJSON. Then when splitting the NDJSON by segment, it chooses a single
sequence per segment by selecting the longest sequence with the latest
accession. The latest accession is represented by the biggest integer
after the "EPI" prefix in the segment accession. Future improvements can
add options for specifying segment accessions and/or sequence quality
annotations from Nextclade.

Note I decided to change the schema of the NDJSON to allow us to easily
pick the segment sequences from records and easily add future
improvements such as deduping strains by available segments. This does
mean that we have to reprocess all previous downloads to regenerate the
cache, but we needed to do this anyways to include all segment sequences.