.github/workflows/ingest: Add option to run via aws-batch
When re-running the workflow from scratch, the action ran into
`No space left on device` error.¹ I tried running the workflow locally
and saw that the total disk usage peaked ~25GB which is past the
GH runner limits of 16GB.²

This commit adds the option to run the ingest workflow via aws-batch
so that we can use it as necessary. I have not added any cpu or memory
overrides, so it will just use the defaults set in our job definition.

¹ <https://github.com/nextstrain/seasonal-flu/pull/274#issuecomment-3644191318>
² <https://docs.github.com/en/actions/reference/runners/github-hosted-runners#standard-github-hosted-runners-for-public-repositories>