ingest: add priority to link_gisaid_metadata_and_fasta
This forces the Snakemake workflow to prioritize creating the compressed
NDJSON before decompressing other pairs of XLS/FASTA files. So in a run
of the workflow in GH Action with 4 CPUs:

1. all 4 CPUs are used to decompress 4 pairs of XLS/FASTA
2. when the one of the pairs is decompressed, the workflow immediately
moves on to create the compressed NDJSON
3. once the compressed NDJSON exists, the decompressed XLS/FASTA pair is
automatically deleted since they are marked as Snakemake temp files
4. the workflow then decompresses the next XLS/FASTA pair

This means no more than 4 pairs of decompressed XLS/FASTA files should
be present at the same time.