ingest: keep gisaid.ndjson uncompressed
Follow up to <https://github.com/nextstrain/seasonal-flu/pull/277/files#r2621286544>

Partially reverts 98b34c97274f1c25d3c85d451ad3cb1c6c50c173 to
keep the intermediate data/gisaid.ndjson file uncompressed. This is
necessary to upload the file with `upload-to-s3` because the script
does not support uploading compressed files yet. The file size is
reasonable to deal with for now (~8GB), but we should pursue adding
compressed file support to `upload-to-s3` and keep the intermediates
compressed soon!  

Alternatives that I considered but did not pursue:
1. Upload the file with a plain `aws s3 cp` command. We would lose the
additional metadata sha256sum and record count and the abililty to skip
uploads if the files are identical.
2. Decompress the file right before upload. This would require the same
disk space as just keeping the file decompressed, so I decided against
the extra decompression step.