ingest/link-gisaid-metadata-and-fasta: Hardcode expected columns
Hardcode expected columns from the metadata XLS so that we can add or
remove columns as needed to keep the cache NDJSON uniform.

Using this as a work-around for the validation error from `augur curate`
as I'm dealing with the lack of columns for originating lab, submitting
lab, and authors.¹ Eventually we should make this process more robust. 

¹ <https://github.com/nextstrain/seasonal-flu/issues/272>