.github/workflow/upload: trigger avian-flu independently
Trigger avian-flu independently after ingest completes so that any
errors that arise does not affect the rest of the seasonal-flu
workflows.

Note that this does duplicate the `gh workflow run` command that
currently points to a WIP branch and trial name. We'll need to
remember to clean up at the same time as ingest. I think that's ~fine
since these will likely stay the same after that update and it's better
than the alternative of making this even more complicated with creating
yet another reusable workflow/composite action.

I had considered removing the trigger from ingest, but I can see the
use case of needing to update curations specific to avian-flu and only
wanting to trigger avian-flu workflows.